+++ src/Accounts/Accounts/Account/ConnectAzureRmAccount.cs
@@ -339,8 +339,7 @@ namespace Microsoft.Azure.Commands.Profile
 
             if (ParameterSetName.Equals(UserWithCredentialParameterSet))
             {
-                string warningMessage = "Authentication with a username and password at the command line is strongly discouraged. Use one of the recommended authentication methods based on your requirements. For additional information, visit https://go.microsoft.com/fwlink/?linkid=2276971.";
-                WriteWarning(warningMessage);
+                WriteWarning(Resources.UsernamePasswordDiscourageWarningMessage);
             }
 
             if (MyInvocation.BoundParameters.ContainsKey(nameof(Subscription)))